Gradle and dependency update
This is a slight cleanup for Gradle. The Gradle wrapper is updated from 2.7 to 3.4 for build performance improvements, bug fixes, and more. Also moves towards a more modern way to apply plugins, updates the shadow and license plugins, and switches to using jcenter() for more secure and faster dependency retrieval, switches the Sponge repository to HTTPS for security purposes, and removes the Sonatype repository as we do not use it. Then finally updates the mongo-java-driver to latest and the HikariCP dependency to what the latest Sponge versions are using.