Clean up usage of Optionals
Optionals are not really intended to be used as fields or parameters, but more as a return type. Nulls are fine to work with and using these as Optionals just adds extra overhead for the GC and it needs to wrap and unwrap Optionals as well, which is more unneccesary work.