Changes the CommandSource in QuerySession to not be an Optional
We were rarely even checking if it was present anyway, so there was going to be errors if it was not there, and we never created one without a command source, so we just had unnecessary boxing and unboxing. If its planned to have a QuerySession without a command source in the future this could be rethought and we could add the necessary checks instead.